data-vis/requirements.txt
pandas
seaborn
numpy
matplotlib
jupyter
cassandra-driver
jupyterlab
plotly
scikit-learn